requirements.txt
can-isotp==2.0.7
dearpygui==2.0.0
msgpack==1.1.0
packaging==25.0
pyserial==3.5
python-can==4.5.0
typing_extensions==4.14.0
wrapt==1.17.2
